Global variables, theme changes to both themes Dark theme fixes
