+++ src/components/Buttons/LanguageButton/LanguageButton.module.scss
@@ -1,4 +1,4 @@
 .languageButton {
-  border-bottom: 1px solid rgba(45, 171, 103, 0.5);
+  border-bottom: 1px solid var(--primary-color-50);
   padding: 1rem 0;
 }
