+++ src/components/ProfileCard/Badge/Badge.module.scss
@@ -7,7 +7,7 @@
     height: 1.5rem;
     display: block;
     path{
-      fill: white;
+      fill: var(--contrast-light-color);
     }
   }
 }
