+++ src/styles/muiExtentions/components.tsx
@@ -1,0 +1,7 @@
+declare module '@mui/material/Button' {
+  interface ButtonPropsColorOverrides {
+    accent: true
+  }
+}
+
+export {}
