+++ src/styles/muiExtentions/palette.tsx
@@ -1,0 +1,32 @@
+declare module '@mui/material/styles' {
+  interface Palette {
+    accent: {
+      main: string
+    }
+    primaryLight: {
+      main: string
+    }
+    constantDark: {
+      main: string
+    }
+    constantLight: {
+      main: string
+    }
+  }
+  interface PaletteOptions {
+    accent?: {
+      main: string
+    }
+    primaryLight?: {
+      main: string
+    }
+    constantDark?: {
+      main: string
+    }
+    constantLight?: {
+      main: string
+    }
+  }
+}
+
+export {}
