+++ src/styles/palette/constant.tsx
@@ -1,0 +1,25 @@
+import { type PaletteOptions } from '@mui/material'
+
+export const constantColors: PaletteOptions = {
+  primary: {
+    main: '#2EAB67',
+  },
+  secondary: {
+    main: '#F9C22E'
+  },
+  error: {
+    main: '#FF36AB'
+  },
+  primaryLight: {
+    main: 'rgba(45, 171, 103, 0.5)'
+  },
+  accent: {
+    main: '#F55D3E'
+  },
+  constantDark: {
+    main: '#0d1b2a'
+  },
+  constantLight: {
+    main: '#D4E2F2'
+  }
+}
