+++ src/styles/shadows.tsx
@@ -1,0 +1,11 @@
+import { type Shadows } from '@mui/material'
+
+export const shadows: Shadows = [
+  'none',
+  `0px 24px 48px rgba(40, 6, 40, 0.1), 0px 12.15px 20.925px rgba(40, 6, 40, 0.0675), 
+  0px 4.8px 7.8px rgba(40, 6, 40, 0.05), 0px 1.05px 2.775px rgba(40, 6, 40, 0.0325)`,
+  `0px 24px 48px rgba(40, 6, 40, 0.1), 0px 12.15px 20.925px rgba(40, 6, 40, 0.0675), 
+  0px 4.8px 7.8px rgba(40, 6, 40, 0.05), 0px 1.05px 2.775px rgba(40, 6, 40, 0.0325)`,
+  'none', 'none', 'none', 'none', 'none', 'none', 'none', 'none', 'none', 'none', 'none',
+  'none', 'none', 'none', 'none', 'none', 'none', 'none', 'none', 'none', 'none', 'none'
+]
