+++ src/styles/typography.tsx
@@ -1,0 +1,28 @@
+import { type TypographyOptions } from '@mui/material/styles/createTypography'
+
+export const typography: TypographyOptions = {
+  fontFamily: 'Lato, sans-serif',
+  h1: {
+    fontFamily: 'Cabin, sans-serif',
+    fontWeight: 600,
+    fontSize: '1.5rem'
+  },
+  h2: {
+    fontFamily: 'Cabin, sans-serif',
+    fontWeight: 600,
+    fontSize: '1.25rem'
+  },
+  body1: {
+    fontSize: '1rem',
+    lineHeight: '130%'
+  },
+  body2: {
+    fontSize: '1rem',
+    lineHeight: '130%',
+    opacity: '60%'
+  },
+  subtitle1: {
+    fontSize: '0.75rem',
+    lineHeight: '130%'
+  }
+}
